Add a Solana rug Blink application
* Objective

This commit objective is about adding a new Solana blink with all the
learnings from the first week (SPL tokens &NFT minting).

* Why

It was a strategic decision to add this as a weekend project,
recognizing the team's integral role in reviewing the concepts for the
week.

* How

This commit adds the required GET and POST endpoints for a Solana Blink
in a vercelapp, performing all the operations needed to mint the NFT
for the user when the Blink is "triggered".